Fix ARM32 cross-compilation type casting in fileinfo_linux.go
Cast timespec seconds and nanoseconds to int64 for time.Unix()
to support 32-bit ARM builds (Raspberry Pi 1-3).

+++ smb/vfs/fileinfo_linux.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package vfs
 
 import (
@@ -20,9 +22,9 @@ func CompatStat(stat os.FileInfo) (Stat, bool) {
 		Uid:     sysStat.Uid,
 		Gid:     sysStat.Gid,
 		Nlink:   uint32(sysStat.Nlink),
-		Atime:   time.Unix(sysStat.Atim.Sec, sysStat.Atim.Nsec),
-		Mtime:   time.Unix(sysStat.Mtim.Sec, sysStat.Mtim.Nsec),
-		Ctime:   time.Unix(sysStat.Ctim.Sec, sysStat.Ctim.Nsec),
+		Atime:   time.Unix(int64(sysStat.Atim.Sec), int64(sysStat.Atim.Nsec)),
+		Mtime:   time.Unix(int64(sysStat.Mtim.Sec), int64(sysStat.Mtim.Nsec)),
+		Ctime:   time.Unix(int64(sysStat.Ctim.Sec), int64(sysStat.Ctim.Nsec)),
 		Btime:   time.Unix(0, 0),
 	}
